refactor(core): share the one default session name builder
SonarQube flagged 15 lines repeated between app-shell and orchestra-canvas.
The bodies of generateDefaultSessionName() were byte-identical; only the
JSDoc example date differed.

defaultSessionName() now lives in frontend core. Both chat and canvas
already import that lib and it sits below both, so the existing canvas to
chat direction is untouched and no new edge appears. frontend ui was the
alternative and was rejected: it holds Floating-UI presentational
primitives, not string logic. Precedent for a plain function in
lib/services is picker-workspace-scope.util.ts and rpc-call.util.ts.

Behaviour is identical, with now defaulting to new Date(). Both components
stay standalone and OnPush.

The 41-line block SonarQube reports in file-settings-keys.spec.ts is left
alone. That spec restates the lane table on purpose, and both files say so:
platform-core is a leaf and cannot import the real SKILL_LANE_DEFAULTS, so
the cross-lib equality is asserted in skills-synthesis-rpc.handlers.spec.ts
instead. The assertions there check the production table against the
restated one. Importing the constant would turn each into expect(x).toBe(x)
and stop detecting a dropped lane or a changed default.
