+++ libs/frontend/core/src/lib/services/index.ts
@@ -44,6 +44,7 @@ export {
   type CommandSuggestion,
 } from './command-discovery.facade';
 export { pickerWorkspaceScope } from './picker-workspace-scope.util';
+export { defaultSessionName } from './default-session-name.util';
 export { setIfChanged } from './idempotent-setters';
 export { LazyViewService } from './lazy-view.service';
 export { BootStatusService } from './boot-status.service';
